[Completes #74964104, #74964146] Create a new morsel should land on Add Title screen. When in drafts, create new morsel should push add title to the stack.

+++ Morsel/Morsel/Source/ViewControllers/MRSLModalCommentsViewController.m
@@ -231,7 +231,7 @@ NSFetchedResultsControllerDelegate>
 #pragma mark - UITableViewDataSource
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    int count = ([_comments count] + ((_previousCommentsAvailable) ? 1 : 0));
+    NSInteger count = ([_comments count] + ((_previousCommentsAvailable) ? 1 : 0));
     return count;
 }
